Added class abstractWeatherContainer.
It is a class containing:
- Wind
- Visibility
- list of clouds
- list of weather conditions
- vertical visibility
